+++ server/drizzle/0008_action_policy_browser_enabled.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "action_policy" ADD COLUMN "browser_enabled" boolean;